graph works, added toggle button to change between time and distance and made a pop up info for graph

+++ app/src/main/java/com/example/sbma_project/database/Run.kt
@@ -23,8 +23,8 @@ data class Run(
     val routePath : List<LatLng>?,
     val rating : Int?,
     val notes : String?,
-    val speed: Int?,
-    val distance: Int?,
+    val speed: Double?,
+    val distance: Double?,
     val steps: Int?,
     val stepLength: Double?,
